src/app/components/summary/summary.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { SummaryComponent } from './summary.component';
import { RouterTestingModule } from '@angular/router/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { HomeComponent } from '../home/home.component';

describe('SummaryComponent', () => {
  let component: SummaryComponent;
  let fixture: ComponentFixture<SummaryComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ SummaryComponent ],
      imports: [
        RouterTestingModule.withRoutes(
          [{path: 'home', component: HomeComponent}]
        ),
        HttpClientTestingModule,
      ],
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(SummaryComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it(`should have as title 'Ingredients Summary`, () => {
    const app = fixture.componentInstance;
    expect(app.title).toEqual('Ingredients Summary');
  });

  it('should render title', () => {
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.highlight-card span').textContent).toContain('Ingredients Summary');
  });

  it('should render summary title', () => {
    component.iTitle = 'Milk';
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.container h3').textContent).toString();
  });

  it('should exist "totalNutrition"', function () {
    expect(component.totalNutrition).toBeDefined();
  });
});
